# platform-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/permission/alluser": {
            "get": {
                "description": "PermissionAllUser Api 查询资源分配给所有人的权限情况",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "PermissionAllUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "资源ID",
                        "name": "resource_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "资源类型",
                        "name": "resource_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PermissionListRes"
                        }
                    }
                }
            },
            "put": {
                "description": "PutPermissionAllUser Api 修改资源分配给所有人的权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "PutPermissionAllUser",
                "parameters": [
                    {
                        "description": "Put Permission AllUser request body",
                        "name": "PutPermissionAllUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ResourcePermissionPutReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/permission/role": {
            "get": {
                "description": "PermissionRole Api 查询资源分配给指定角色的情况",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "PermissionRole",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "资源ID",
                        "name": "resource_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "资源类型",
                        "name": "resource_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RoleResourcePermissionRes"
                        }
                    }
                }
            },
            "put": {
                "description": "PutPermissionRole Api 修改资源分配给指定用户的权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "PutPermissionRole",
                "parameters": [
                    {
                        "description": "Put Permission User request body",
                        "name": "ResourcePermissionPutUserReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ResourcePermissionPutUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/permission/user": {
            "get": {
                "description": "PermissionUser Api 查询资源分配给指定人的权限情况",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "PermissionUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "资源ID",
                        "name": "resource_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "资源类型",
                        "name": "resource_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserResourcePermissionRes"
                        }
                    }
                }
            },
            "put": {
                "description": "PutPermissionUser Api 修改资源分配给指定用户的权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "PutPermissionUser",
                "parameters": [
                    {
                        "description": "Put Permission User request body",
                        "name": "ResourcePermissionPutUserReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ResourcePermissionPutUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/policy/create": {
            "put": {
                "description": "PolicyCreate Api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Policy"
                ],
                "summary": "PolicyCreate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Create Policy request body",
                        "name": "PolicyCreate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PolicyCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/policy/delete": {
            "delete": {
                "description": "PolicyDelete Api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Policy"
                ],
                "summary": "PolicyDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Del Policy request body",
                        "name": "PolicyDelete",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PolicyDelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/policy/update": {
            "post": {
                "description": "PolicyUpdate Api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Policy"
                ],
                "summary": "PolicyUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update Policy request body",
                        "name": "PolicyUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PolicyUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.Permission": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "子资源，可能为空",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Permission"
                    }
                },
                "key": {
                    "$ref": "#/definitions/enum.PermissionKey"
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                }
            }
        },
        "dto.PermissionListRes": {
            "type": "object",
            "properties": {
                "menu": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/enum.PermissionKey"
                    }
                },
                "permission_tree": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Permission"
                    }
                }
            }
        },
        "dto.PolicyCreateReq": {
            "type": "object",
            "required": [
                "associate_id",
                "associate_type",
                "policy_name",
                "policy_type"
            ],
            "properties": {
                "associate_id": {
                    "type": "integer"
                },
                "associate_type": {
                    "$ref": "#/definitions/enum.AssociateType"
                },
                "cron_str": {
                    "type": "string"
                },
                "execute_type": {
                    "type": "integer"
                },
                "policy_name": {
                    "type": "string"
                },
                "policy_type": {
                    "$ref": "#/definitions/enum.PolicyType"
                },
                "single_exec_time": {
                    "type": "integer"
                }
            }
        },
        "dto.PolicyDelReq": {
            "type": "object",
            "required": [
                "policy_id"
            ],
            "properties": {
                "policy_id": {
                    "type": "integer"
                }
            }
        },
        "dto.PolicyUpdateReq": {
            "type": "object",
            "required": [
                "associate_id",
                "associate_type",
                "policy_id",
                "policy_type"
            ],
            "properties": {
                "associate_id": {
                    "type": "integer"
                },
                "associate_type": {
                    "$ref": "#/definitions/enum.AssociateType"
                },
                "cron_str": {
                    "type": "string"
                },
                "policy_id": {
                    "type": "integer"
                },
                "policy_name": {
                    "type": "string"
                },
                "policy_type": {
                    "$ref": "#/definitions/enum.PolicyType"
                },
                "single_exec_time": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "dto.ResourcePermissionPutReq": {
            "type": "object",
            "required": [
                "resource_id",
                "resource_type"
            ],
            "properties": {
                "enable": {
                    "type": "boolean"
                },
                "permission_key": {
                    "$ref": "#/definitions/enum.PermissionKey"
                },
                "resource_id": {
                    "type": "integer"
                },
                "resource_type": {
                    "$ref": "#/definitions/enum.ResourceType"
                }
            }
        },
        "dto.ResourcePermissionPutUserReq": {
            "type": "object",
            "required": [
                "resource_id",
                "resource_type"
            ],
            "properties": {
                "enable": {
                    "type": "boolean"
                },
                "permission_key": {
                    "$ref": "#/definitions/enum.PermissionKey"
                },
                "resource_id": {
                    "type": "integer"
                },
                "resource_type": {
                    "$ref": "#/definitions/enum.ResourceType"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.RolePermission": {
            "type": "object",
            "properties": {
                "keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/enum.PermissionKey"
                    }
                },
                "role_id": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
        "dto.RoleResourcePermissionRes": {
            "type": "object",
            "properties": {
                "role_permission": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RolePermission"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UserPermission": {
            "type": "object",
            "properties": {
                "keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/enum.PermissionKey"
                    }
                },
                "user_display_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "dto.UserResourcePermissionRes": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "user_permission": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserPermission"
                    }
                }
            }
        },
...
        "enum.AssociateType": {
            "type": "integer",
            "enum": [
                1,
                2
            ],
            "x-enum-varnames": [
                "AssociateTypeSingle",
                "AssociateTypeGroup"
            ]
        },
...
        "enum.PermissionKey": {
            "type": "string",
            "enum": [
                "view",
                "edit",
                "delete",
                "poweron",
                "poweroff",
                "permissionmgmt"
            ],
            "x-enum-varnames": [
                "PermissionKeyView",
                "PermissionKeyEdit",
                "PermissionKeyDelete",
                "PermissionKeyPowerOn",
                "PermissionKeyPowerOff",
                "PermissionKeyPermissionMgmt"
            ]
        },
        "enum.PolicyType": {
            "type": "string",
            "enum": [
                "on",
                "off"
            ],
            "x-enum-varnames": [
                "PolicyTypeOn",
                "PolicyTypeOff"
            ]
        },
        "enum.ResourceType": {
            "type": "string",
            "enum": [
                "role",
                "device",
                "device_group",
                "user"
            ],
            "x-enum-varnames": [
                "ResourceTypeRole",
                "ResourceTypeDevice",
                "ResourceTypeDeviceGroup",
                "ResourceTypeUser"
            ]
        },
...
    }
}